can't not using setRole

+++ src/main/java/com/main/laptop_world/Repository/RoleRepository.java
@@ -4,7 +4,9 @@ import com.main.laptop_world.Entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Integer> {
     Role findByName(String name);
 }
